+++ api/production_settings-dist.py
@@ -1,4 +1,8 @@
 from .settings import *
+import sentry_sdk
+from sentry_sdk.integrations.django import DjangoIntegration
+
+DEBUG = False
 
 DATABASES = {
     'default': {
@@ -14,3 +18,13 @@ DATABASES = {
 ALLOWED_HOSTS = []
 
 STATIC_ROOT = os.path.join(BASE_DIR, 'static/')
+
+CORS_ORIGIN_WHITELIST = []
+
+SENTRY_SDK_KEY = ''
+
+# Set sentry sdk key here
+sentry_sdk.init(
+    dsn=SENTRY_SDK_KEY,
+    integrations=[DjangoIntegration()]
+)